# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HEADSETTING Key Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSession {
  loggedIn: boolean;
  /** @nullable */
  username?: string | null;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface DashboardStats {
  totalKeys: number;
  activeKeys: number;
  expiredKeys: number;
  bannedKeys: number;
  totalDevices: number;
  todayActivations: number;
}

export type LicenseKeyStatus =
  (typeof LicenseKeyStatus)[keyof typeof LicenseKeyStatus];

export const LicenseKeyStatus = {
  active: "active",
  expired: "expired",
  banned: "banned",
} as const;

export interface LicenseKey {
  id: number;
  keyString: string;
  status: LicenseKeyStatus;
  maxDevices: number;
  deviceCount: number;
  expiredDate: string;
  createdAt: string;
}

export interface KeysResponse {
  keys: LicenseKey[];
}

export interface GenerateKeyBody {
  days: number;
  maxDevices: number;
}

export interface EditKeyBody {
  maxDevices: number;
  expiredDate: string;
}

export interface Device {
  id: number;
  deviceId: string;
  /** @nullable */
  deviceName?: string | null;
  /** @nullable */
  androidVersion?: string | null;
  firstActivated: string;
  lastCheckin: string;
}

export interface DevicesResponse {
  devices: Device[];
}

export interface ActivityLog {
  id: number;
  action: string;
  /** @nullable */
  keyString?: string | null;
  /** @nullable */
  deviceId?: string | null;
  /** @nullable */
  detail?: string | null;
  timestamp: string;
}

export interface LogsResponse {
  logs: ActivityLog[];
}

export interface PublicCheckResponse {
  valid: boolean;
  message: string;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  expiredDate?: string | null;
  /** @nullable */
  maxDevices?: number | null;
}

export interface PublicActivateBody {
  key: string;
  device_id: string;
  device_name?: string;
  android_version?: string;
}

export interface PublicActivateResponse {
  success: boolean;
  message: string;
  /** @nullable */
  alreadyRegistered?: boolean | null;
  /** @nullable */
  key?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  expiredDate?: string | null;
  /** @nullable */
  daysRemaining?: number | null;
  /** @nullable */
  maxDevices?: number | null;
  /** @nullable */
  deviceCount?: number | null;
  /** @nullable */
  deviceId?: string | null;
  /** @nullable */
  deviceName?: string | null;
}

export interface VerifyBody {
  key: string;
  device_id?: string;
  device_name?: string;
  android_version?: string;
}

export interface VerifyResponse {
  status: string;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  expired_date?: string | null;
}

export type PublicCheckKeyParams = {
  key: string;
  device_id?: string;
  device_name?: string;
  android_version?: string;
};