# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HEADSETTING Key Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Admin authentication
  - name: keys
    description: License key management
  - name: logs
    description: Activity logs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /stats:
    get:
      operationId: getStats
      tags: [keys]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /keys:
    get:
      operationId: getKeys
      tags: [keys]
      summary: List all license keys
      responses:
        "200":
          description: List of keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeysResponse"

  /keys/generate:
    post:
      operationId: generateKey
      tags: [keys]
      summary: Generate a new license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateKeyBody"
      responses:
        "201":
          description: Key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"

  /keys/{id}:
    put:
      operationId: editKey
      tags: [keys]
      summary: Edit key expiry/device limit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditKeyBody"
      responses:
        "200":
          description: Key updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"
    delete:
      operationId: deleteKey
      tags: [keys]
      summary: Delete a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Key deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /keys/{id}/ban:
    post:
      operationId: banKey
      tags: [keys]
      summary: Ban a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Key banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"

  /keys/{id}/unban:
    post:
      operationId: unbanKey
      tags: [keys]
      summary: Unban a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Key unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"

  /keys/{id}/devices:
    get:
      operationId: getKeyDevices
      tags: [keys]
      summary: Get devices registered to a key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Devices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevicesResponse"

  /keys/{id}/devices/reset:
    post:
      operationId: resetKeyDevices
      tags: [keys]
      summary: Reset all devices for a key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Devices reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /logs:
    get:
      operationId: getLogs
      tags: [logs]
      summary: Get activity logs
      responses:
        "200":
          description: Activity logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"

  /public/check:
    get:
      operationId: publicCheckKey
      tags: [keys]
      summary: Validasi key dari script eksternal (tanpa login)
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
        - name: device_id
          in: query
          required: false
          schema:
            type: string
        - name: device_name
          in: query
          required: false
          schema:
            type: string
        - name: android_version
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Hasil pengecekan key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCheckResponse"

  /public/activate:
    post:
      operationId: publicActivateKey
      tags: [keys]
      summary: Aktivasi device ke key (tanpa login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicActivateBody"
      responses:
        "201":
          description: Aktivasi berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicActivateResponse"
        "200":
          description: Hasil pengecekan (gagal / sudah terdaftar)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicActivateResponse"

  /verify:
    post:
      operationId: verifyKey
      tags: [keys]
      summary: Verifikasi key dari script shell
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyBody"
      responses:
        "200":
          description: Hasil verifikasi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"

  /verify/heartbeat:
    post:
      operationId: verifyHeartbeat
      tags: [keys]
      summary: Heartbeat ping dari script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyBody"
      responses:
        "200":
          description: Hasil heartbeat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    AdminSession:
      type: object
      properties:
        loggedIn:
          type: boolean
        username:
          type: ["string", "null"]
      required: [loggedIn]

    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required: [success]

    DashboardStats:
      type: object
      properties:
        totalKeys:
          type: integer
        activeKeys:
          type: integer
        expiredKeys:
          type: integer
        bannedKeys:
          type: integer
        totalDevices:
          type: integer
        todayActivations:
          type: integer
      required: [totalKeys, activeKeys, expiredKeys, bannedKeys, totalDevices, todayActivations]

    LicenseKeyStatus:
      type: string
      enum: [active, expired, banned]

    LicenseKey:
      type: object
      properties:
        id:
          type: integer
        keyString:
          type: string
        status:
          $ref: "#/components/schemas/LicenseKeyStatus"
        maxDevices:
          type: integer
        deviceCount:
          type: integer
        expiredDate:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, keyString, status, maxDevices, deviceCount, expiredDate, createdAt]

    KeysResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/LicenseKey"
      required: [keys]

    GenerateKeyBody:
      type: object
      properties:
        days:
          type: integer
        maxDevices:
          type: integer
      required: [days, maxDevices]

    EditKeyBody:
      type: object
      properties:
        maxDevices:
          type: integer
        expiredDate:
          type: string
      required: [maxDevices, expiredDate]

    Device:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        deviceName:
          type: ["string", "null"]
        androidVersion:
          type: ["string", "null"]
        firstActivated:
          type: string
          format: date-time
        lastCheckin:
          type: string
          format: date-time
      required: [id, deviceId, firstActivated, lastCheckin]

    DevicesResponse:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/Device"
      required: [devices]

    ActivityLog:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        keyString:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        detail:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
      required: [id, action, timestamp]

    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
      required: [logs]

    PublicCheckResponse:
      type: object
      properties:
        valid:
          type: boolean
        message:
          type: string
        key:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        expiredDate:
          type: ["string", "null"]
        maxDevices:
          type: ["integer", "null"]
      required: [valid, message]

    PublicActivateBody:
      type: object
      properties:
        key:
          type: string
        device_id:
          type: string
        device_name:
          type: string
        android_version:
          type: string
      required: [key, device_id]

    PublicActivateResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        alreadyRegistered:
          type: ["boolean", "null"]
        key:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        expiredDate:
          type: ["string", "null"]
        daysRemaining:
          type: ["integer", "null"]
        maxDevices:
          type: ["integer", "null"]
        deviceCount:
          type: ["integer", "null"]
        deviceId:
          type: ["string", "null"]
        deviceName:
          type: ["string", "null"]
      required: [success, message]

    VerifyBody:
      type: object
      properties:
        key:
          type: string
        device_id:
          type: string
        device_name:
          type: string
        android_version:
          type: string
      required: [key]

    VerifyResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: ["string", "null"]
        expired_date:
          type: ["string", "null"]
      required: [status]